requirements.txt
python-telegram-bot==20.8
openai==1.12.0
python-dotenv==1.0.1
schedule==1.2.1
pytz==2024.1
pytest==8.0.2 